Refactor checkin screen to use partials
Each partial only has the data passed to it - sometimes this is a lot of data (like the list of racers being passed to `main_checkin_table.php`), sometimes it's little to none (e.g. passing `partition_label` and `partition_label_pl` to `edit_racer_modal.php`.
Javascript which is tied directly to a specific modal is included in the partial for that modal.

This helps with separation of concerns by ensuring that the partials are only responsible for displaying data - business logic (such as permission checks) or data fetching happens in the "controller" which in this case is the root-level `checkin.php`.
