+++ website/partials/checkin/barcode_settings_modal.php
@@ -1,0 +1,15 @@
+<div id="barcode_settings_modal" class="modal_dialog hidden block_buttons">
+    <form>
+        <h2>Barcode Responses</h2>
+        <input id="barcode-handling-locate" name="barcode-handling" type="radio" value="locate"/>
+        <label for="barcode-handling-locate">Locate racer</label>
+        <input id="barcode-handling-checkin" name="barcode-handling" type="radio" value="checkin"/>
+        <label for="barcode-handling-checkin">Check in racer</label>
+        <input id="barcode-handling-racer" name="barcode-handling" type="radio" value="racer-photo"/>
+        <label for="barcode-handling-racer">Capture racer photo</label>
+        <input id="barcode-handling-car" name="barcode-handling" type="radio" value="car-photo"/>
+        <label for="barcode-handling-car">Capture car photo</label>
+
+        <input type="submit" value="Close"/>
+    </form>
+</div>
