+++ website/partials/checkin/qrcode_settings_modal.php
@@ -1,0 +1,11 @@
+<div id="qrcode_settings_modal" class="modal_dialog wide_modal hidden block_buttons">
+    <form id="mobile-checkin-form">
+        <h2>Mobile Check-In</h2>
+        <div id="mobile-checkin-qrcode" style="width: 256px; margin-left: 122px;"></div>
+        <div id="mobile-checkin-title" style="text-align: center; font-size: 1.5em; font-weight: bold; margin-bottom: 25px; margin-top: 10px;"></div>
+        <input id="mobile-checkin-url" name="mobile-checkin-url" type="text" />
+
+        <a id="mcheckin-link" class="button_link" href="mcheckin.php">Mobile Check-In &gt;</a>
+        <input type="button" value="Close" onclick="close_modal('#qrcode_settings_modal');"/>
+    </form>
+</div>
